rename: `Followers`를 `@pages/`에서 `@layouts`으로 옮김 #66
